Lab9/webServiceA/server/requirements.txt
asgiref==3.2.7
Django==3.0.4
django-background-task==0.1.8
django-background-tasks==1.2.5
django-compat==1.0.15
django-cors-headers==3.2.1
djangorestframework==3.11.0
pika==1.1.0
pkg-resources==0.0.0
pytz==2019.3
six==1.14.0
sqlparse==0.3.1
